Add Github Actions workflows
This commit adds Github Actions workflows for building the XR
Fundamentals codelab. The workflows are defined in the
following files:

- .github/workflows/xr-fundamentals-part1-build-app.yml
- .github/workflows/xr-fundamentals-part2-build-app.yml
- .github/workflows/xr-fundamentals-start-build-app.yml

These workflows will be triggered on push and pull request
events to the main branch. They can also trigger manually
with `workflow_dispatch`. They will build the app for each
part of the codelab.